Card generator cleanup
